refactor: extract the data-dir role lock into crates/single-instance
The flock/LockFileEx guard moves out of the desktop's platform module so
oximux serve can contend for per-data-dir roles (the schedule ticker) with
identical semantics. The desktop keeps the GUI-only halves — the lock
filename and the window-activation paths — and re-exports the lock API
from the crate, so its callers are unchanged.
